creates 2 new files: users_model.js and users_controller.js and creates the relevant endpoint on app.js file. Creates a new app.test.js file to keep all the tests on a single file and deletes the obsolete test files
