+++ app.js
@@ -1,10 +1,16 @@
 const express = require("express");
+
 const { getTopics } = require("./controllers/topics_controller");
+
+const { getUsers } = require("./controllers/users_controller");
+
 const { getEndpoints } = require("./controllers/endpoints_controller");
+
 const {
   getArticlesById,
   updateArticle,
 } = require("./controllers/articles_controller");
+
 const {
   getCommentsByArticleId,
   addCommentsForArticleId,
@@ -18,6 +24,7 @@ app.use(express.json());
 app.get("/api/topics", getTopics);
 app.get("/api/articles/:article_id", getArticlesById);
 app.get("/api/articles/:article_id/comments", getCommentsByArticleId);
+app.get("/api/users", getUsers);
 app.get("/api", getEndpoints);
 app.post("/api/articles/:article_id/comments", addCommentsForArticleId);
 app.patch("/api/articles/:article_id", updateArticle);
