+++ controllers/users_controller.js
@@ -1,0 +1,10 @@
+const { fetchUsers } = require("../models/users_model");
+
+exports.getUsers = async (req, res, next) => {
+  try {
+    const users = await fetchUsers();
+    res.status(200).send({ users });
+  } catch (err) {
+    next(err);
+  }
+};
