+++ models/comments_model.js
@@ -5,7 +5,6 @@ exports.selectCommentsByArticleId = async (article_id) => {
     "SELECT * FROM comments WHERE article_id = $1 ORDER BY created_at DESC;",
     [article_id]
   );
-  //
   return comments.rows;
 };
 
@@ -31,7 +30,7 @@ exports.insertComment = async (username, body, article_id) => {
 
 exports.deleteCommentByCommentId = async (comment_id) => {
   const deletedComment = await db.query(
-    "DELETE FROM comments WHERE comment_id = $1 RETURNING *",
+    "DELETE FROM comments WHERE comment_id = $1",
     [comment_id]
   );
   return deletedComment.rowCount;
