+++ models/topics_model.js
@@ -1,6 +1,6 @@
 const db = require("../db/connection");
 
 exports.fetchTopics = async () => {
-  const result = await db.query(`SELECT * FROM topics`);
+  const result = await db.query(`SELECT * FROM topics;`);
   return result.rows;
 };
