+++ models/users_model.js
@@ -1,0 +1,6 @@
+const db = require("../db/connection");
+
+exports.fetchUsers = async () => {
+  const result = await db.query(`SELECT * FROM users;`);
+  return result.rows;
+};
